admin and department password reset

+++ src/main/java/com/ayushsingh/cacmp_backend/services/serviceimpl/DepartmentServiceImpl.java
@@ -1,12 +1,12 @@
 package com.ayushsingh.cacmp_backend.services.serviceimpl;
 
 import com.ayushsingh.cacmp_backend.models.dtos.departmentDtos.DepartmentDetailsDto;
+import com.ayushsingh.cacmp_backend.models.dtos.departmentDtos.DepartmentPasswordChangeDto;
 import com.ayushsingh.cacmp_backend.models.dtos.departmentDtos.DepartmentRegisterDto;
 import com.ayushsingh.cacmp_backend.models.entities.Department;
 import com.ayushsingh.cacmp_backend.models.projections.department.DepartmentNameProjection;
 import com.ayushsingh.cacmp_backend.models.roles.DepartmentRole;
 import com.ayushsingh.cacmp_backend.repository.entities.CategoryRepository;
-import com.ayushsingh.cacmp_backend.repository.entities.ComplaintRepository;
 import com.ayushsingh.cacmp_backend.repository.entities.DepartmentRepository;
 import com.ayushsingh.cacmp_backend.repository.entities.PollRepository;
 import com.ayushsingh.cacmp_backend.services.DepartmentRoleService;
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-@Slf4j
 @RequiredArgsConstructor
 @Service
 public class DepartmentServiceImpl implements DepartmentService {
@@ -34,8 +33,9 @@ public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
     private final PasswordEncoder passwordEncoder;
     private final DepartmentRoleService departmentRoleService;
-   private final CategoryRepository categoryRepository;
+    private final CategoryRepository categoryRepository;
     private final ModelMapper modelMapper;
+
     @Override
     public Boolean isDepartmentCredentialsPresent(String username) {
         return departmentRepository.existsByUsername(username);
@@ -43,20 +43,19 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public String registerDepartment(DepartmentRegisterDto departmentDto) {
-        String username=departmentDto.getUsername();
+        String username = departmentDto.getUsername();
         Boolean isDepartmentCredentialsPresent = isDepartmentCredentialsPresent(username);
-        if(isDepartmentCredentialsPresent){
-            throw new RuntimeException("Department with username: "+username+" already exists");
-        }
-        else{
-            Department department=new Department();
+        if (isDepartmentCredentialsPresent) {
+            throw new RuntimeException("Department with username: " + username + " already exists");
+        } else {
+            Department department = new Department();
             department.setUsername(username);
             department.setPassword(passwordEncoder.encode(departmentDto.getPassword()));
             department.setDepartmentName(departmentDto.getDepartmentName());
             department.setDepartmentObjective(departmentDto.getDepartmentObjective());
-            department=departmentRepository.save(department);
-            Set<DepartmentRole> roles=new HashSet<>();
-            for(String role:departmentDto.getRoles()){
+            department = departmentRepository.save(department);
+            Set<DepartmentRole> roles = new HashSet<>();
+            for (String role : departmentDto.getRoles()) {
                 roles.add(departmentRoleService.getDepartmentRoleByRoleName(role));
             }
             department.setRoles(roles);
@@ -73,9 +72,9 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public List<DepartmentDetailsDto> listAllDepartments() {
-        List<Department> departments=departmentRepository.findAll();
-        List<DepartmentDetailsDto> departmentDetailsDtos=departments.stream().map(department -> {
-            return this.modelMapper.map(department,DepartmentDetailsDto.class);
+        List<Department> departments = departmentRepository.findAll();
+        List<DepartmentDetailsDto> departmentDetailsDtos = departments.stream().map(department -> {
+            return this.modelMapper.map(department, DepartmentDetailsDto.class);
         }).toList();
         return departmentDetailsDtos;
     }
@@ -83,10 +82,13 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Transactional
     @Override
     public void deleteDepartment(String departmentToken) {
-        Long count=categoryRepository.getCategoryCountByDepartmentUsername(departmentToken);
-        Long countPolls=pollRepository.countByDepartmentToken(departmentToken);
-        if(count!=0){
+        Long count = categoryRepository.getCategoryCountByDepartmentUsername(departmentToken);
+        Long countPolls = pollRepository.countByDepartmentToken(departmentToken);
+        if (count != 0) {
             throw new ApiException("Department cannot be deleted because there are categories associated with it");
+        }
+        if (countPolls != 0) {
+            throw new ApiException("Department cannot be deleted because there are active polls associated with it");
         }
         departmentRepository.deleteByDeptToken(departmentToken);
     }
@@ -94,26 +96,24 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     public DepartmentDetailsDto getDepartment(String departmentToken) {
         Optional<Department> departmentOptional = departmentRepository.findByDeptToken(departmentToken);
-        if(departmentOptional.isEmpty()){
-            throw new ApiException("Department with department token: "+departmentToken+" does not exist");
-        }
-        else{
-            return this.modelMapper.map(departmentOptional.get(),DepartmentDetailsDto.class);
+        if (departmentOptional.isEmpty()) {
+            throw new ApiException("Department with department token: " + departmentToken + " does not exist");
+        } else {
+            return this.modelMapper.map(departmentOptional.get(), DepartmentDetailsDto.class);
         }
     }
 
     @Override
     public String updateDepartment(DepartmentDetailsDto departmentDetailsDto) {
-        String token=departmentDetailsDto.getDeptToken();
+        String token = departmentDetailsDto.getDeptToken();
         Optional<Department> departmentOptional = departmentRepository.findByDeptToken(token);
-        if(departmentOptional.isEmpty()){
-            throw new ApiException("Department with department token: "+token+" does not exist");
-        }
-        else{
-            Department department=departmentOptional.get();
+        if (departmentOptional.isEmpty()) {
+            throw new ApiException("Department with department token: " + token + " does not exist");
+        } else {
+            Department department = departmentOptional.get();
             department.setDepartmentName(departmentDetailsDto.getDepartmentName());
             department.setDepartmentObjective(departmentDetailsDto.getDepartmentObjective());
-            department=departmentRepository.save(department);
+            department = departmentRepository.save(department);
             return department.getDeptToken();
         }
     }
@@ -123,5 +123,19 @@ public class DepartmentServiceImpl implements DepartmentService {
         return departmentRepository.findALlDepartmentNames();
     }
 
+    @Override
+    public String updateDepartmentPassword(DepartmentPasswordChangeDto departmentPasswordChangeDto) {
+        Optional<Department> departmentOptional = departmentRepository
+                .findByDeptToken(departmentPasswordChangeDto.getDepartmentToken());
+        if (departmentOptional.isEmpty()) {
+            throw new ApiException("Department with department token: "
+                    + departmentPasswordChangeDto.getDepartmentToken() + " does not exist");
+        } else {
+            Department department = departmentOptional.get();
+            department.setPassword(passwordEncoder.encode(departmentPasswordChangeDto.getPassword()));
+            department = departmentRepository.save(department);
+            return department.getDeptToken();
+        }
+    }
 
 }
